split IW test to common
